+++ src/Pipedrive.net/Models/Common/SimplePicture.cs
@@ -11,9 +11,9 @@ namespace Pipedrive
         public string Url { get; set; }
 
         [JsonProperty("width")]
-        public int Width { get; set; }
+        public long Width { get; set; }
 
         [JsonProperty("height")]
-        public int Height { get; set; }
+        public long Height { get; set; }
     }
 }
