+++ src/Pipedrive.net/Models/Response/Pipeline.cs
@@ -14,7 +14,7 @@ namespace Pipedrive
         public string UrlTitle { get; set; }
 
         [JsonProperty("order_nr")]
-        public int OrderNr { get; set; }
+        public long OrderNr { get; set; }
 
         [JsonProperty("active")]
         public bool Active { get; set; }
